Fixing many compilation errors for windows. Compiler now passes. Only missing some implementations nw.

+++ src/Types.h
@@ -4,7 +4,8 @@ typedef uint8_t                 byte;
 typedef uint8_t                 uint8;
 typedef uint16_t                uint16;
 typedef uint32_t                uint32;
-typedef unsigned long long int  uint64;
+//typedef unsigned long long int  uint64;
+typedef uint64_t uint64;
 
 #if !uint
     typedef uint32          uint;
